split read and write providers
